+++ src/main/java/com/github/olegushak/FTT/command/CommandName.java
@@ -7,10 +7,11 @@ public enum CommandName {
     START("/start"),
     STOP("/stop"),
     HELP("/help"),
-    FIND_FLIGHT("find ticket"),
+    FIND_FLIGHT("find flight"),
     SUBSCRIBE_TICKET("subscribe to ticket"),
     NO_COMMAND("no command"),
     STAT("/stat"),
+    SEARCH_AIRPORT("/search airport"),
     SHARE_LOCATION("/share location"),
     UPDATE_LOCALISATIONS("/update_localisations"),
     UPDATE_AIRPORTS("/update_airports");