+++ src/main/java/com/github/olegushak/FTT/command/searchCommand/SearchCommand.java
@@ -1,0 +1,7 @@
+package com.github.olegushak.FTT.command.searchCommand;
+
+import org.telegram.telegrambots.meta.api.objects.Update;
+
+public interface SearchCommand {
+    void execute(Update update );
+}