+++ src/main/java/com/github/olegushak/FTT/utils/EmojiConst.java
@@ -1,0 +1,11 @@
+package com.github.olegushak.FTT.utils;
+
+import com.vdurmont.emoji.EmojiParser;
+
+public class EmojiConst {
+    public final static String DEP_AIRPLANE = EmojiParser.parseToUnicode("\uD83D\uDEEB");
+    public final static String ARR_AIRPLANE = EmojiParser.parseToUnicode("\uD83D\uDEEC");
+    public final static String DEP_DATE = EmojiParser.parseToUnicode("\uD83D\uDDD3");
+    public final static String RETURN_DATE = EmojiParser.parseToUnicode("\uD83D\uDCC6");
+
+}