+++ RELEASE_NOTES.md
@@ -1,5 +1,17 @@
 # Release notes
 
+## 0.6.0-SNAPSHOT
+
+* Installed Elasticsearch container.
+* implemented flight ticket search form.
+* implemented fuzzy city search .
+* created buttons for search form.
+
+## 0.5.0-SNAPSHOT
+
+* added the localisations and airports DB
+* added the "Menu" button to the bot
+
 ## 0.4.0-SNAPSHOT
 
 * added the possibility to get the all available tickets for one way flight